МУЗ.ИИ/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
spacy>=3.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
click>=8.1.0
httpx>=0.25.0